requirements.txt
playwright==1.48.0
customtkinter==5.2.2
openpyxl==3.1.5
pandas==2.2.3
ddddocr==1.4.11
opencv-python-headless==4.10.0.84
numpy==1.26.4
cryptography==43.0.3
Pillow==10.4.0
